use Decimal type

+++ src/domain/dto.rs
@@ -1,5 +1,6 @@
 use super::client::{Country, Document};
 use chrono::NaiveDate;
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
@@ -14,11 +15,11 @@ pub struct NewClientBody {
 #[derive(Deserialize, Serialize)]
 pub struct NewCreditTransactionBody {
     pub client_id: Uuid,
-    pub credit_amount: f64,
+    pub credit_amount: Decimal,
 }
 
 #[derive(Deserialize, Serialize)]
 pub struct NewDebitTransactionBody {
     pub client_id: Uuid,
-    pub debit_amount: f64,
+    pub debit_amount: Decimal,
 }
